IMPALA-9335 (part 2): Fix rebased KRPC to compile
This patch applies various fixes to Impala and to the copied Kudu
source code in be/src/kudu/* to allow everything to compile.

Some highlights of the changes made:
- Various Kudu files were removed from compilation due to issues like
  relying on libraries that Impala does not provide. The linking of
  some executable is also changed for similar reasons.
- The Kudu Cache implementation changed to support unique_ptr,
  allowing us to remove various uses of MakeScopeExitTrigger.
- Some flags that have a DEFINE in both Kudu and Impala are modified
  to change one of the DEFINEs to a DECLARE.

This patch was in part based on the patches that were applied the last
time we rebased the Kudu code in IMPALA-7006, and I ensured that all
changes from those commits that are still relevant were included here.

I also went through all commits that have been applied to the
be/src/kudu directory since the last rebase and ensured that all
relevant changes from those are included here.

Testing:
- Passed an exhaustive DEBUG build and a core ASAN build.

Change-Id: I1eb4caf927c729109426fb50a28b5e15d6ac46cb
Reviewed-on: http://gerrit.cloudera.org:8080/15144
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>