IMPALA-9342: Membership updates should only remove quiescing nodes from the blacklist
Currently, the ClusterMembershipMgr will remove a node from the
blacklist whenever there is an "update" for a backend from the statestore.
Updates are typically restricted to updates about the quiescing status
of a node. The ClusterMembershipMgr should un-blacklist quiescing nodes
since quiescing nodes are not part of any executor groups and will
eventually be removed from the cluster membership. Thus, there is no
reason they need to remain on the blacklist.

However, other updates to a backend (e.g. updates that are not related
to the quiescing status of a node) should not cause that node to be
un-blacklisted. Doing so could cause a node to be un-blacklisted, but
not added back to any executor groups, creating a state where a node is
part of the cluster membership, but not part of any executor groups (or
the blacklist).

This patch fixes the aforementioned issue by only un-blacklisting an
updated node in ClusterMembershipMgr::UpdateMembership when the node
starts quiescing. Added some DCHECKs to ensure the consistency of the
blacklist and the list of executor groups.

Testing:
* Ran core tests
* Ran test_executor_groups.p, test_restart_services.py,
  and test_blacklist.py with --exploration_strategy=exhaustive
  locally

Change-Id: Id062e51df86315ac214d30db882736dbb7948a77
Reviewed-on: http://gerrit.cloudera.org:8080/15137
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>