IMPALA-9262: Bump statestore_heartbeat_frequency_ms in test_kill_impalad_with_running_queries
CustomClusterTestSuite sets statestore_heartbeat_frequency_ms to 50,
overriding the default value of 1000. This means that if a node does not
respond to heartbeats for 500 milliseconds, it will time out and be
removed from the cluster (vs the default of 10 seconds).

A low value for statestore_heartbeat_frequency_ms is problematic for
test_kill_impalad_with_running_queries because if a node is removed from
the cluster membership, it is removed from the blacklist as well. The
test asserts that "Blacklisted Executors" shows up in the runtime
profile of a query immediately after running a query that causes a node
to be blacklisted. Thus, there is a race condition between running the
test query vs. the node being removed from the cluster membership.
Increasing the value of statestore_heartbeat_frequency_ms should
significantly reduce the chances of such a race.

Testing:
* Ran test_kill_impalad_with_running_queries locally
* Not actually able to re-produce the flakiness locally

Change-Id: I84e884efab35649b63db1a7a3b8c49b95b0b4648
Reviewed-on: http://gerrit.cloudera.org:8080/15131
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>