IMPALA-8690 (prep 1): Copy Kudu cache code to be/src/util/cache
The data cache uses Kudu's cache code in be/src/kudu/util.
To make it easy to make Impala specific changes (such as
introducing new cache eviction algorithms), this copies the
code into Impala's utilities directory.

To make the review easier, this is a simple copy of the files
with a minimal CMakeLists.txt and RAT changes. This does not link
the resulting library into anything.

A subsequent change will modify the code to be usable by
Impala. To test, I verified that this compiles.

Change-Id: I0e7b8b852b56613803dea125456fa9e8b3736d76
Reviewed-on: http://gerrit.cloudera.org:8080/15169
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>
Tested-by: Joe McDonnell <joemcdonnell@cloudera.com>