IMPALA-8852: Skip short-circuit config check for dedicated coordinator
ImpalaD should not abort when running as dedicated coodinator and DataNode is
not available on the host. This change adds a condition to skip the short-
circuit socket path directory checks when ImpalaD is started with
'is_executor=false' flag.

Testing:
 - Added test to JniFrontendTest.java to verify the short-circuit directory
check is skipped if ImpalaD is started as dedicated coordinator mode.
 - Manually tested the appearance of the warning message with:
start-impala-cluster.py --num_coordinators 1 --use_exclusive_coordinators true

Change-Id: I373d4037f4cee203322a398b77b75810ba708bb5
Reviewed-on: http://gerrit.cloudera.org:8080/15173
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>