IMPALA-9359: recover from corrupt kerberos ccache
This is a clean cherry-pick of KUDU-3050. The original commit
message is below.

KUDU-3050: recover from corrupt kerberos ccache

This handles two failure modes:
* krb5_cc_start_seq_get() can fail if the kerberos credential cache gets
  corrupted on disk, e.g. is truncated.
* the renewal can fail to find a credential in the credential cache,
  either if it is missing or the renewal thread hits an error while
  reading through credentials.

Also add some additional logging and limit the max backoff time
to make it easier to debug other kinds of renewal errors.

The test triggers a pre-existing memory leak bug in some older
Kerberos libraries. Added a suppression for leak sanitizer
to ClientNegotiation::CheckGSSAPI() to suppress it.

Test:
Add a test that exercises the recovery logic after truncating
the credential cache. The test failed before this change.

Change-Id: I86567f16816d1c6729679398ce56296744cb30c9
Reviewed-on: http://gerrit.cloudera.org:8080/15407
Reviewed-by: Thomas Tauber-Marshall <tmarshall@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>