IMPALA-8870: Bump up Guava to 28.1-jre and set DISABLE_SENTRY to true
This patch bumps up the version of Guava libraries from 14.0.1 to
28.1-jre. Due to some changes in Guava's API's, we modify the call
sites accordingly.

Moreover, in order to instruct the Java classes under the directory of
$IMPALA_HOME/common/yarn-extras to use the new Guava libraries, we
explicitly added a dependency in the corresponding pom.xml file.

On the other hand, we set DISABLE_SENTRY to true regardless of
$USE_CDP_HIVE since Sentry's Guava version has not been bumped up yet
and thus run-sentry-service.sh cannot be successfully executed. Recall
that by setting DISABLE_SENTRY to true we also disable every
Sentry-related test, which is fine from now on since Impala 3.4 was
recently branched. The plan is to drop support for Sentry in the Impala
4 line.

Change-Id: I9690a926953a8d3c3872277680b4be0551546c68
Reviewed-on: http://gerrit.cloudera.org:8080/15214
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>