IMPALA-9399: optimise RuntimeProfile::ToThrift()
The big win here is to preallocate the vector<TRuntimeProfileNode>
to the right size, to avoid copying the TRuntimeProfileNode objects
when reallocating the vector. The JIRA description on IMPALA-9399
explains why this happens.

Also various simplification/modernisation of code to avoid
unnecessary copies of vectors and redundant lock acquisitions.

Perf:
Added a microbenchmark that shows a 2.75x speedup of
RuntimeProfile::ToThrift() on a synthetic profile.

Change-Id: I1f85ef005ae3b793515b223d4c04bbd898336065
Reviewed-on: http://gerrit.cloudera.org:8080/15251
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>