IMPALA-9400: Initial Ozone support
Ozone is a block-based filesystem store and is similar to HDFS in
many aspects. This patch adds initial support for reading data stored
on Ozone and writing data to Ozone. It adds Ozone to the whitelist of
Impala writable filesystems and adds the necessary Ozone client jars
to the Impala classpath.

Initially, Ozone will be treated as a remote storage service, similar to
how Impala treats S3 and ABFS. Ozone will typically be deployed on a
separate storage cluster, so all reads will be remote.

Since remote HDFS and Ozone are separate storage systems, with distinct
performance characteristics, they are treated as separate systems for
disk scheduling purposes.

Rather than updating to a new CDP GBN with the added Ozone client jars,
I just modified the existing GBN to include the new jars in the Maven
repo.

Testing:
* Manually tested that connecting to an external Ozone cluster works and
  that Impala can create an Ozone-backed table, write to the
  table, read from the table, and drop the table
* Ran core tests with and without USE_CDP_HIVE=true

Change-Id: I2544ff61e8767f061253a62a01ce2b7dfc384a99
Reviewed-on: http://gerrit.cloudera.org:8080/15334
Reviewed-by: Sahil Takiar <stakiar@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>