IMPALA-9402: Relax bootstrap condition for PostgreSQL permission tweaks
bin/bootstrap_system.sh was looking for a strict combination of
OS platform / pg_hba.conf location / default authN method
when setting up the default Impala build environment.

This combo proved to be too strict for CentOS 7.4 running with
PostreSQL 9.2, and failed to set up this environment correctly,
which lead to a failure in the dataload phase.

The patch fixes the permission tweak for this combination by relaxing
the initial condition where the tweak may be applied.

Change-Id: I31cbee6c22f9c27b1ff2b9200bb356be14429f39
Reviewed-on: http://gerrit.cloudera.org:8080/15243
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>