IMPALA-9395: fix duplicate broadcast SetFilter() calls
For broadcast join filters, we can get duplicate filters
- one set from the coordinator and one set locally.
The intent is that the first-arriving filter is used
(it's exposed to consumers via SetFilter()) and the
later filters are ignored.

This should be guaranteed by checking HasFilter()
in the same critical section as SetFilter(). I
checked all code paths in runtime-filter-bank.cc
that call SetFilter() for broadcast filters and
fixed the single exception.

Testing:
I wasn't able to reproduce this issue locally. Ran
runtime filter tests to check for regressions.

Change-Id: I95d0620c4dbb5e4066702db48442cebee7389f5a
Reviewed-on: http://gerrit.cloudera.org:8080/15242
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>