IMPALA-9351: Change path in TestCreateTableLikeFileOrc
There are some builds where AnalyzeDDLTest.TestCreateTableLikeFileOrc
failed due to non-existing path. The problem was with the complex
table check:
/test-warehouse/functional_orc_def.db/complextypes_fileformat/000000_0'

As Joe suggested, this commit changes the path to another complex
table, "complextypestbl", which is loaded by Impala and not written
by Hive, so the location should always be the same.

Tests:
 - Run the changed test.

Change-Id: I94ac0f09a738e6c3edb068b036c1d50dc9861a7f
Reviewed-on: http://gerrit.cloudera.org:8080/15266
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>