IMPALA-9401: primitive include-what-you-use script and mappings
This is a cleaned up version of the script I used to run
include-what-you-use on the Impala codebase.

The helper script assumes you have built IWYU and
have a Kudu source checkout, and can then run IWYU
on the entire codebase.

Some mappings files are used to improve the quality of
the IWYU output. There are still incorrect recommendations
made, but this is sufficient to fix many common issues.

Change-Id: Iaf5f9ba865313afb0c581e6482514ef7f1c65367
Reviewed-on: http://gerrit.cloudera.org:8080/15552
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>